修复关注列表bug

+++ common/response.go
@@ -45,7 +45,7 @@ type Video struct {
 type User struct {
 	FollowCount   int64  `json:"follow_count"`   // 关注总数
 	FollowerCount int64  `json:"follower_count"` // 粉丝总数
-	ID            int64  `json:"id"`             // 用户id
+	ID            uint   `json:"id"`             // 用户id
 	IsFollow      bool   `json:"is_follow"`      // true-已关注，false-未关注
 	Name          string `json:"name"`           // 用户名称
 }
